pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "qraft"
version = "0.1.0"
description = "Lightweight SQL templating and orchestration tool"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Ravidhu Dissanayake", email = "ravidhu.dissa@gmail.com" }]
keywords = ["sql", "templating", "orchestration", "data-engineering", "dbt", "etl", "dag"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Rust",
    "Topic :: Database",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "duckdb>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
postgres = ["psycopg[binary]>=3.1"]
mysql = ["pymysql>=1.1"]
trino = ["trino>=0.328"]
all-engines = ["psycopg[binary]>=3.1", "pymysql>=1.1", "trino>=0.328"]

[project.urls]
Homepage = "https://github.com/ravidhu/qraft"
Repository = "https://github.com/ravidhu/qraft"
Documentation = "https://github.com/ravidhu/qraft/tree/main/docs"
Issues = "https://github.com/ravidhu/qraft/issues"
Changelog = "https://github.com/ravidhu/qraft/blob/main/CHANGELOG.md"

[project.scripts]
qraft = "qraft.cli:main"

[tool.maturin]
# Pointe vers le crate Rust
manifest-path = "rust/Cargo.toml"
# Le module compilé est placé dans le package Python
module-name = "qraft._core"
# Source Python
python-source = "python"
# Features PyO3
features = ["pyo3/extension-module"]
# Include catalog static files in the wheel
include = ["python/qraft/catalog/static/**"]

[tool.pytest.ini_options]
testpaths = ["python/qraft/tests"]
pythonpath = ["python"]

[tool.mypy]
python_version = "3.11"
mypy_path = "python"

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
